+++ packages/extension/src/adapter/cdp-drivers/index.ts
@@ -2,4 +2,5 @@
 // T1.7a / T1.7b / T1.7c 逐步迁入。
 
 export { runDateRangeDriverCDP } from "./daterange.js";
-// runCascaderDriverCDP / runTimePickerDriverCDP 在 T1.7b / T1.7c 加入
+export { runCascaderDriverCDP } from "./cascader.js";
+// runTimePickerDriverCDP 在 T1.7c 加入
